+++ packages/vue/src/index.ts
@@ -1,8 +1,8 @@
-export { BlockComponent } from './Block';
-export { LinkComponent, ActiveLink } from './Link';
-export { ReactiveResponse } from './interface';
-export { CuriPluginOptions } from './plugin';
+export { BlockComponent } from "./Block";
+export { LinkComponent } from "./Link";
+export { ReactiveResponse } from "./interface";
+export { CuriPluginOptions } from "./plugin";
 
-import CuriPlugin from './plugin';
+import CuriPlugin from "./plugin";
 
 export { CuriPlugin };